+++ models/Facility.py
@@ -2,4 +2,14 @@ class Facility:
 
     def __init__(self):
         self.name = None
-        self.generatedID = 1  # change once database is queried
+        self.state = None
+        # self.generatedID = self.state + "_" + self.name
+
+    def getDict(self):
+        return {"_id": self.generatedID,
+                "name": self.name,
+                "state": self.state
+                }
+
+    def getGeneratedID(self):
+        return self.generatedID
